Fix binomial memoization to use long

+++ Binomial.java
@@ -17,12 +17,12 @@ public class Binomial {
 	
 	
 	// Computes the Binomial function, efficiently
-	public static int binomial(int n, int k) {
+	public static long binomial(int n, int k) {
 
 	    if (k > n) return 0;
     	if (k == 0 || n == 0) return 1;
 
-		int [][] memo = new int [n+1][k+1]; 
+		long [][] memo = new long [n+1][k+1]; 
 
 		for (int i = 0; i<memo.length;i++)
 		{
@@ -35,7 +35,7 @@ public class Binomial {
 		return binomial(n, k, memo);
 	}
 
-	private static int binomial(int n, int k, int[][] memo) {
+	private static long binomial(int n, int k, long[][] memo) {
 		if (memo[n][k] != -1) {
 			return memo[n][k];
 		}
